// client/web/angular-guitar-helper/generated-sources/openapi/chords-service-openapi/model/chord.ts
/**
 * Guitara Chords Service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GuitarStringState } from './guitarStringState';
import { GuitarBarrePushed } from './guitarBarrePushed';
import { GuitarPositionPushed } from './guitarPositionPushed';


export interface Chord { 
    id?: string;
    name: string;
    firstFretReference?: Chord.FirstFretReferenceEnum;
    mutedOpenStrings: Set<GuitarStringState>;
    positionsPushed: Set<GuitarPositionPushed>;
    barreFrets: Set<GuitarBarrePushed>;
}
export namespace Chord {
    export const FirstFretReferenceEnum = {
        _1: '1',
        _2: '2',
        _3: '3',
        _4: '4',
        _5: '5',
        _6: '6',
        _7: '7',
        _8: '8',
        _9: '9',
        _10: '10',
        _11: '11',
        _12: '12',
        _13: '13',
        _14: '14',
        _15: '15',
        _16: '16',
        _17: '17',
        _18: '18',
        _19: '19',
        _20: '20',
        _21: '21',
        _22: '22'
    } as const;
    export type FirstFretReferenceEnum = typeof FirstFretReferenceEnum[keyof typeof FirstFretReferenceEnum];
}