// client/web/angular-guitar-helper/generated-sources/openapi/chords-service-openapi/model/guitarBarrePushed.ts
/**
 * Guitara Chords Service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GuitarBarrePushed { 
    fret?: GuitarBarrePushed.FretEnum;
    startNote?: GuitarBarrePushed.StartNoteEnum;
    endNote?: GuitarBarrePushed.EndNoteEnum;
    fingerNumber?: GuitarBarrePushed.FingerNumberEnum;
}
export namespace GuitarBarrePushed {
    export const FretEnum = {
        _1: '1',
        _2: '2',
        _3: '3',
        _4: '4',
        _5: '5',
        _6: '6',
        _7: '7',
        _8: '8',
        _9: '9',
        _10: '10',
        _11: '11',
        _12: '12',
        _13: '13',
        _14: '14',
        _15: '15',
        _16: '16',
        _17: '17',
        _18: '18',
        _19: '19',
        _20: '20',
        _21: '21',
        _22: '22'
    } as const;
    export type FretEnum = typeof FretEnum[keyof typeof FretEnum];
    export const StartNoteEnum = {
        E: 'e',
        A: 'A',
        D: 'D',
        G: 'G',
        B: 'B',
        E2: 'E'
    } as const;
    export type StartNoteEnum = typeof StartNoteEnum[keyof typeof StartNoteEnum];
    export const EndNoteEnum = {
        E: 'e',
        A: 'A',
        D: 'D',
        G: 'G',
        B: 'B',
        E2: 'E'
    } as const;
    export type EndNoteEnum = typeof EndNoteEnum[keyof typeof EndNoteEnum];
    export const FingerNumberEnum = {
        _1: '1',
        _2: '2',
        _3: '3',
        _4: '4'
    } as const;
    export type FingerNumberEnum = typeof FingerNumberEnum[keyof typeof FingerNumberEnum];
}