// client/web/angular-guitar-helper/generated-sources/openapi/chords-service-openapi/model/guitarStringState.ts
/**
 * Guitara Chords Service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GuitarStringState { 
    guitarString?: GuitarStringState.GuitarStringEnum;
    openCloseState?: GuitarStringState.OpenCloseStateEnum;
}
export namespace GuitarStringState {
    export const GuitarStringEnum = {
        E: 'e',
        A: 'A',
        D: 'D',
        G: 'G',
        B: 'B',
        E2: 'E'
    } as const;
    export type GuitarStringEnum = typeof GuitarStringEnum[keyof typeof GuitarStringEnum];
    export const OpenCloseStateEnum = {
        _0: '0',
        _1: '1'
    } as const;
    export type OpenCloseStateEnum = typeof OpenCloseStateEnum[keyof typeof OpenCloseStateEnum];
}